+++ demo/src/app/components/+carousel/demos/config/config.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+import { CarouselConfig } from 'ngx-bootstrap/carousel';
+
+@Component({
+  selector: 'demo-carousel-config',
+  templateUrl: './config.html',
+  providers: [{provide: CarouselConfig, useValue: {interval: 1500, noPause: true}}]
+})
+export class DemoCarouselConfigComponent {
+}