+++ demo/src/app/components/+tabs/demos/basic/basic.ts
@@ -1,0 +1,25 @@
+import { Component, ViewChild } from '@angular/core';
+import { TabsetComponent } from 'ngx-bootstrap';
+
+@Component({
+  selector: 'demo-tabs-basic',
+  templateUrl: './basic.html'
+})
+export class DemoTabsBasicComponent {
+
+  @ViewChild('staticTabs') staticTabs: TabsetComponent;
+
+  public alertMe(): void {
+    setTimeout(function (): void {
+      alert('You\'ve selected the alert tab!');
+    });
+  }
+
+  selectTab(tab_id: number) {
+      this.staticTabs.tabs[tab_id].active = true;
+  }
+
+  disableEnable() {
+    this.staticTabs.tabs[2].disabled = ! this.staticTabs.tabs[2].disabled
+  }
+}