+++ demo/src/polyfills.ts
@@ -1,4 +1,4 @@
-// This file includes polyfills needed by Angular 2 and is loaded before
+// This file includes polyfills needed by Angular and is loaded before
 // the app. You can add your own extra polyfills to this file.
 import 'ts-helpers';
 import 'core-js/es6/symbol';