+++ demo/src/typings.d.ts
@@ -6,8 +6,8 @@ declare const ENV:string;
 // google code-prettify
 declare const PR:any;
 
-declare const require:any;
-declare const global:any;
+// declare const require:any;
+// declare const global:any;
 
 declare module jasmine {
   interface Matchers {