+++ src/datepicker/date-formatter.ts
@@ -1,4 +1,4 @@
-import * as moment from 'moment';
+import moment from 'moment';
 
 export class DateFormatter {
   public format(date:Date, format:string):string {