+++ src/modal/index.ts
@@ -1,4 +1,4 @@
-export * from './modal-backdrop.component';
-export * from './modal-options.class';
-export * from './modal.component';
+export { ModalBackdropComponent, ModalBackdropOptions } from './modal-backdrop.component';
+export { ModalOptions } from './modal-options.class';
+export { ModalDirective } from './modal.component';
 export { ModalModule } from './modal.module';