+++ src/popover/popover.directive.ts
@@ -1,14 +1,6 @@
 import {
-  Directive,
-  Input,
-  Output,
-  EventEmitter,
-  OnInit,
-  OnDestroy,
-  Renderer,
-  ElementRef,
-  TemplateRef,
-  ViewContainerRef
+  Directive, Input, Output, EventEmitter, OnInit, OnDestroy, Renderer,
+  ElementRef, TemplateRef, ViewContainerRef
 } from '@angular/core';
 import { PopoverConfig } from './popover.config';
 import { ComponentLoaderFactory, ComponentLoader } from '../component-loader';
@@ -110,6 +102,8 @@ export class PopoverDirective implements OnInit, OnDestroy {
       this._outsideClickListener = this._renderer.listenGlobal('document', 'mousedown',
         ($event: MouseEvent) => this.onMouseDown($event.target));
     }
+
+    this.isOpen = true;
   }
 
   /**
@@ -119,6 +113,7 @@ export class PopoverDirective implements OnInit, OnDestroy {
   public hide(): void {
     if (this.isOpen) {
       this._popover.hide();
+      this.isOpen = false;
     }
     if (this._outsideClickListener) {
       this._outsideClickListener();
@@ -153,7 +148,7 @@ export class PopoverDirective implements OnInit, OnDestroy {
     if (!this._getContainerNativeElement()) return;
     if (this._elementRef.nativeElement === target) return; // if popover toggle element was clicked
     if (this._getContainerNativeElement().contains(target)) return; // if clicked on popover content
-    
+
     this.hide();
   }